Сделал абстрактный компонент и подключил через него другие компоненты
