+++ src/components/board.js
@@ -1,6 +1,4 @@
-import {
-  createElement
-} from '../utils';
+import AbstractComponent from './abstract-component';
 
 const createMainFilmTemplate = () => {
   return (
@@ -23,25 +21,9 @@ const createMainFilmTemplate = () => {
   );
 };
 
-class Board {
-  constructor() {
-    this._element = null;
-  }
-
+class Board extends AbstractComponent {
   getTemplate() {
     return createMainFilmTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
