+++ src/components/footer-stats.js
@@ -1,33 +1,15 @@
+import AbstractComponent from './abstract-component';
 import {
   cardCount
-} from '../const';
-import {
-  createElement
-} from '../utils';
+} from '../utils/const';
 
 const createFilmsCountTemplate = () => {
   return `<p>${cardCount.FILMS_CARDS_COUNT} movies inside</p>`;
 };
 
-class FilmsCount {
-  constructor() {
-    this._element = null;
-  }
-
+class FilmsCount extends AbstractComponent {
   getTemplate() {
     return createFilmsCountTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
