+++ src/components/no-films.js
@@ -1,30 +1,12 @@
-import {
-  createElement
-} from '../utils';
+import AbstractComponent from './abstract-component';
 
 const createNoFilmsTemplate = () => {
   return `<h2 class="films-list__title">There are no movies in our database</h2>`;
 };
 
-class NoFilms {
-  constructor() {
-    this._element = null;
-  }
-
+class NoFilms extends AbstractComponent {
   getTemplate() {
     return createNoFilmsTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
