+++ src/components/show-more-button.js
@@ -1,6 +1,4 @@
-import {
-  createElement
-} from '../utils';
+import AbstractComponent from './abstract-component';
 
 const createShowMoreButtonTemplate = () => {
   return (
@@ -8,25 +6,9 @@ const createShowMoreButtonTemplate = () => {
   );
 };
 
-class ShowMoreButton {
-  constructor() {
-    this._element = null;
-  }
-
+class ShowMoreButton extends AbstractComponent {
   getTemplate() {
     return createShowMoreButtonTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
