+++ src/utils/common.js
@@ -1,0 +1,14 @@
+const switchElem = (arr, className) => {
+  arr.forEach((elem) => {
+    elem.addEventListener(`click`, () => {
+      arr.forEach((item) => {
+        item.classList.remove(`${className}--active`);
+      });
+      elem.classList.add(`${className}--active`);
+    });
+  });
+};
+
+export {
+  switchElem,
+};
