+++ src/utils/render.js
@@ -20,19 +20,7 @@ const render = (container, element, place) => {
   }
 };
 
-const switchElem = (arr, className) => {
-  arr.forEach((elem) => {
-    elem.addEventListener(`click`, () => {
-      arr.forEach((item) => {
-        item.classList.remove(`${className}--active`);
-      });
-      elem.classList.add(`${className}--active`);
-    });
-  });
-};
-
 export {
   createElement,
   render,
-  switchElem,
 };
